+++ packages/go_router/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 15.2.0
+
+- Adds new top level `onEnter` callback with access to current and next route states.
+- Deprecates top level `redirect` in favor of `onEnter`.
+
 ## 15.1.1
 
 - Adds missing `caseSensitive` to `GoRouteData.$route`.
@@ -16,7 +21,7 @@
 ## 14.8.1
 
 - Secured canPop method for the lack of matches in routerDelegate's configuration.
- 
+
 ## 14.8.0
 
 - Adds `preload` parameter to `StatefulShellBranchData.$branch`.